options: Include more checks for color/no-color
Using crayon::num_ansi_colors would be nice, but we can't force the
stream it's checking, and it assumes no-color in knitr regardless of
output stream, meaning we'd get no color in vignette stderr output.

Instead, add some more checks / overrides that crayon uses, in
particular so we can force color on for CMD check.