outcome: Explicitly record_outcomes()
Instead of switching based on being interactive, add a
record_outcomes() which initalises the outcomes data.frame(), and
switch outcome recording behaviour based on whether this is about.

This also means we can tell the difference between outcomes not being
recorded and no tests ran. We can't put out a message yet though, since
we thoroughly break CMD check in doing so.