outcome: Don't blat history on subsequent record_outcomes()
If a ut_with_report() is nested in a non-interactive script, e.g, The
second call to record_outcomes() shouldn't blat the history of the
first. The history should be preserved & ut_with_report() should no-op,
showing a history at the very end.

The other option would be treating ut_with_report() as a subtest by
having a stack of outcome data.frames, but this seems unlikely to be
what you want.

This invalidates most of the ut_with_report() tests, so rewrite them
with run_script()

Also remove stray assign('errors') that should have been removed
earlier.