format_traceback: Unify error-handling in try_catch_traceback()
We have several places now where we try in various ways to tryCatch,
and get a traceback at the end. Unify all of these under
try_catch_traceback().

To do this we lose the ability to hide Traceback when empty, and
showing "Whilst evaluating" (since now just shows tryCatch), but
neither is a great loss.