outcome: Common line-writing, stop_on_fail-handling
ok_group() wasn't honouring stop_on_fail, fix this by moving the
handling into assign_outcome(), along with the ok / not ok line
writing.

This makes for an uglier stack trace when stop_on_fail is on, but the
stack-trace is unnecessary anyway, so detect this and show a nicer
error message.