ok_group: Summarise groups and tests separately
groups aren't tests, and it doesn't make sense to say a test failed
when it can't ever succeed. Separate the group / test outcomes and
report them separately.

Ideally the plan would only include tests too, and be stable regardless
of group failures. We could use static analysis to try and work this
out, but we'd make a mess if ok()s are in a loop.